feat(2543): bundle optimization
