+++ packages/bodiless-components/src/Editable.tsx
@@ -12,70 +12,18 @@
  * limitations under the License.
  */
 
-import React, { ComponentType as CT, ClipboardEvent } from 'react';
+import React, { ComponentType as CT } from 'react';
 import { flow } from 'lodash';
-import ContentEditable from 'react-contenteditable';
-import { observer } from 'mobx-react-lite';
 import {
   withNode,
-  useNode,
   withNodeKey,
-  useEditContext,
   withChild,
 } from '@bodiless/core';
+import EditableNode from './EditableNode.dev';
 import './Editable.css';
 
-type Props = {
-  placeholder?: string;
-};
-type Data = {
-  text: string;
-};
-
-/** @type {{search: React.DOMAttributes}} */
-const Text = observer((props: Props) => {
-  const { placeholder } = props;
-  const { node } = useNode<Data>();
-  const text = node.data.text || placeholder || '';
-  // eslint-disable-next-line react/no-danger
-  return <span dangerouslySetInnerHTML={{ __html: text }} />;
-});
-const EditableText = observer((props: Props) => {
-  const { node } = useNode<Data>();
-  const onInput = (event: any) => {
-    const newText = event.currentTarget.innerHTML;
-    node.setData({ text: newText });
-  };
-  const pasteAsPlainText = (event: ClipboardEvent<HTMLDivElement>) => {
-    if (event.clipboardData) {
-      event.preventDefault();
-      const text = event!.clipboardData.getData('text/plain');
-      document.execCommand('insertHTML', false, text);
-    }
-  };
-  const { placeholder } = props;
-  const placeholderDataAttr = placeholder || '';
-  const text = node.data.text || '';
-  return (
-    <ContentEditable
-      tagName="span"
-      className="bodiless-inline-editable"
-      onChange={onInput}
-      onPaste={pasteAsPlainText}
-      html={text}
-      data-placeholder={placeholderDataAttr}
-    />
-  );
-});
 const Editable = withNode(
-  observer((props: any) => {
-    const { isEdit } = useEditContext();
-    return isEdit ? (
-      <EditableText {...props} />
-    ) : (
-      <Text {...props} />
-    );
-  }),
+  (props: any) => <EditableNode {...props} />,
 );
 const withPlaceholder = <P extends object> (placeholder?: string) => (Component:CT<P> | string) => {
   const WithPlaceholder = placeholder
