+++ packages/bodiless-components/src/EditableNode.prod.tsx
@@ -1,0 +1,35 @@
+/**
+ * Copyright © 2019 Johnson & Johnson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+
+  Production version of EditableNode optimized for rendering
+
+*/
+
+import React from 'react';
+import { useNode } from '@bodiless/core';
+import { Data, Props } from './EditableNodeCommon';
+
+const EditableNode = (props: Props) => {
+  const { placeholder } = props;
+  const { node } = useNode<Data>();
+  const text = node.data.text || placeholder || '';
+  // eslint-disable-next-line react/no-danger
+  return (
+    <span dangerouslySetInnerHTML={{ __html: text }} />
+  )
+};
+
+export default EditableNode;
