+++ packages/bodiless-components/src/EditableNodeCommon.tsx
@@ -1,0 +1,30 @@
+/**
+ * Copyright © 2019 Johnson & Johnson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+
+  Shared code between production and development versions of EditableNode
+
+*/
+
+type Props = {
+  placeholder?: string;
+};
+
+type Data = {
+  text: string;
+};
+
+// eslint-disable-next-line no-undef
+export { Props, Data };
